scripts/requirements.txt
boto3>=1.26.0
python-dateutil>=2.8.0
psycopg2-binary>=2.9.0
requests>=2.28.0
